# packages/contracts/openapi/control-plane.v1.yaml
openapi: 3.1.0
info:
  title: Private Cloud Control Plane API
  version: 1.0.0
  summary: Provider-neutral virtual-machine lifecycle control plane
  description: |
    Accepts durable asynchronous lifecycle commands and exposes project-scoped read models.
    A `202` response means intent and its operation record committed; it does not mean provider work completed.
  license:
    name: Apache-2.0
    identifier: Apache-2.0
servers:
  - url: https://cloud.example.invalid
    description: Environment-specific Gateway API endpoint
security:
  - bearerAuth: []
...
paths:
...
  /v1/projects/{projectId}:
    parameters:
      - $ref: './components.v1.yaml#/components/parameters/ProjectId'
    get:
      operationId: getProject
      tags: [Projects]
      summary: Read an authorized project
      x-audience: tenant
      x-idempotency: read-only
      x-error-codes: [AUTHENTICATION_REQUIRED, PROJECT_ACCESS_DENIED, PROJECT_NOT_FOUND]
      x-requirements: [SRS-FR-001, SRS-FR-002, SRS-FR-003, SRS-FR-006]
      responses:
        '200':
          description: Project detail
          content:
            application/json:
              schema:
                $ref: './components.v1.yaml#/components/schemas/Project'
        '401': { $ref: './components.v1.yaml#/components/responses/Unauthorized' }
        '403': { $ref: './components.v1.yaml#/components/responses/Forbidden' }
        '404': { $ref: './components.v1.yaml#/components/responses/NotFound' }
  /v1/projects/{projectId}/quota:
    parameters:
      - $ref: './components.v1.yaml#/components/parameters/ProjectId'
    get:
      operationId: getProjectQuota
      tags: [Projects]
      summary: Read project quota limits and usage
      x-audience: tenant
      x-idempotency: read-only
      x-error-codes: [PROJECT_ACCESS_DENIED, PROJECT_NOT_FOUND]
      x-requirements: [SRS-FR-002, SRS-FR-012]
      responses:
        '200':
          description: Project quota
          content:
            application/json:
              schema:
                $ref: './components.v1.yaml#/components/schemas/QuotaSet'
        '401': { $ref: './components.v1.yaml#/components/responses/Unauthorized' }
        '403': { $ref: './components.v1.yaml#/components/responses/Forbidden' }
        '404': { $ref: './components.v1.yaml#/components/responses/NotFound' }
  /v1/projects/{projectId}/catalog/images:
    parameters:
      - $ref: './components.v1.yaml#/components/parameters/ProjectId'
      - $ref: './components.v1.yaml#/components/parameters/Cursor'
      - $ref: './components.v1.yaml#/components/parameters/Limit'
    get:
      operationId: listImages
      tags: [Catalog]
      summary: List enabled provider-neutral images
      x-audience: tenant
      x-idempotency: read-only
      x-error-codes: [VALIDATION_FAILED, PROJECT_ACCESS_DENIED]
      x-requirements: [SRS-FR-002, SRS-FR-007, SRS-FR-008]
      responses:
        '200':
          description: Image page
          content:
            application/json:
              schema:
                $ref: './components.v1.yaml#/components/schemas/ImagePage'
        '400': { $ref: './components.v1.yaml#/components/responses/BadRequest' }
        '401': { $ref: './components.v1.yaml#/components/responses/Unauthorized' }
        '403': { $ref: './components.v1.yaml#/components/responses/Forbidden' }
  /v1/projects/{projectId}/catalog/flavors:
    parameters:
      - $ref: './components.v1.yaml#/components/parameters/ProjectId'
      - $ref: './components.v1.yaml#/components/parameters/Cursor'
      - $ref: './components.v1.yaml#/components/parameters/Limit'
    get:
      operationId: listFlavors
      tags: [Catalog]
      summary: List enabled provider-neutral flavors
      x-audience: tenant
      x-idempotency: read-only
      x-error-codes: [VALIDATION_FAILED, PROJECT_ACCESS_DENIED]
      x-requirements: [SRS-FR-002, SRS-FR-007, SRS-FR-009]
      responses:
        '200':
          description: Flavor page
          content:
            application/json:
              schema:
                $ref: './components.v1.yaml#/components/schemas/FlavorPage'
        '400': { $ref: './components.v1.yaml#/components/responses/BadRequest' }
        '401': { $ref: './components.v1.yaml#/components/responses/Unauthorized' }
        '403': { $ref: './components.v1.yaml#/components/responses/Forbidden' }
  /v1/projects/{projectId}/catalog/networks:
    parameters:
      - $ref: './components.v1.yaml#/components/parameters/ProjectId'
      - $ref: './components.v1.yaml#/components/parameters/Cursor'
      - $ref: './components.v1.yaml#/components/parameters/Limit'
    get:
      operationId: listNetworks
      tags: [Catalog]
      summary: List enabled provider-neutral networks
      x-audience: tenant
      x-idempotency: read-only
      x-error-codes: [VALIDATION_FAILED, PROJECT_ACCESS_DENIED]
      x-requirements: [SRS-FR-002, SRS-FR-007, SRS-FR-010]
      responses:
        '200':
          description: Network page
          content:
            application/json:
              schema:
                $ref: './components.v1.yaml#/components/schemas/NetworkPage'
        '400': { $ref: './components.v1.yaml#/components/responses/BadRequest' }
        '401': { $ref: './components.v1.yaml#/components/responses/Unauthorized' }
        '403': { $ref: './components.v1.yaml#/components/responses/Forbidden' }
...
  /v1/projects/{projectId}/operations/{operationId}:
    parameters:
      - $ref: './components.v1.yaml#/components/parameters/ProjectId'
      - $ref: './components.v1.yaml#/components/parameters/OperationId'
    get:
      operationId: getOperation
      tags: [Operations]
      summary: Read one project operation
      x-audience: tenant
      x-idempotency: read-only
      x-error-codes: [PROJECT_ACCESS_DENIED, OPERATION_NOT_FOUND]
      x-requirements: [SRS-FR-002, SRS-FR-021, SRS-FR-022]
      responses:
        '200':
          description: Operation detail
          content:
            application/json:
              schema:
                $ref: './components.v1.yaml#/components/schemas/Operation'
        '401': { $ref: './components.v1.yaml#/components/responses/Unauthorized' }
        '403': { $ref: './components.v1.yaml#/components/responses/Forbidden' }
        '404': { $ref: './components.v1.yaml#/components/responses/NotFound' }
...
  /v1/admin/operations/{operationId}:
    parameters:
      - $ref: './components.v1.yaml#/components/parameters/OperationId'
    get:
      operationId: getAdministrativeOperation
      tags: [Administration]
      summary: Read one operation with recovery metadata
      x-audience: administrator
      x-idempotency: read-only
      x-error-codes: [ADMIN_REQUIRED, OPERATION_NOT_FOUND]
      x-requirements: [SRS-FR-004, SRS-FR-023]
      responses:
        '200':
          description: Administrative operation detail
          content:
            application/json:
              schema: { $ref: './components.v1.yaml#/components/schemas/AdministrativeOperation' }
        '401': { $ref: './components.v1.yaml#/components/responses/Unauthorized' }
        '403': { $ref: './components.v1.yaml#/components/responses/Forbidden' }
        '404': { $ref: './components.v1.yaml#/components/responses/NotFound' }
  /v1/admin/dead-letters:
    get:
      operationId: listDeadLetters
      tags: [Administration]
      summary: List dead-letter evidence
      x-audience: administrator
      x-idempotency: read-only
      x-error-codes: [VALIDATION_FAILED, ADMIN_REQUIRED]
      x-requirements: [SRS-FR-004, SRS-FR-066]
      parameters:
        - $ref: './components.v1.yaml#/components/parameters/Cursor'
        - $ref: './components.v1.yaml#/components/parameters/Limit'
      responses:
        '200':
          description: Dead-letter page
          content:
            application/json:
              schema: { $ref: './components.v1.yaml#/components/schemas/DeadLetterPage' }
        '400': { $ref: './components.v1.yaml#/components/responses/BadRequest' }
        '401': { $ref: './components.v1.yaml#/components/responses/Unauthorized' }
        '403': { $ref: './components.v1.yaml#/components/responses/Forbidden' }
  /v1/admin/dead-letters/{eventId}/replays:
    parameters:
      - $ref: './components.v1.yaml#/components/parameters/EventId'
    post:
      operationId: replayDeadLetter
      tags: [Administration]
      summary: Accept an attributed dead-letter replay
      x-audience: administrator
      x-idempotency: required
      x-max-body-bytes: 65536
      x-error-codes:
        [ADMIN_REQUIRED, DEAD_LETTER_NOT_FOUND, REPLAY_NOT_ALLOWED, IDEMPOTENCY_CONFLICT]
      x-requirements: [SRS-FR-004, SRS-FR-005, SRS-FR-067, SRS-NFR-019]
      parameters:
        - $ref: './components.v1.yaml#/components/parameters/IdempotencyKey'
        - $ref: './components.v1.yaml#/components/parameters/CorrelationId'
        - $ref: './components.v1.yaml#/components/parameters/Traceparent'
        - $ref: './components.v1.yaml#/components/parameters/Tracestate'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: './components.v1.yaml#/components/schemas/ReplayRequest' }
      responses:
        '202': { $ref: './components.v1.yaml#/components/responses/MutationAccepted' }
        '401': { $ref: './components.v1.yaml#/components/responses/Unauthorized' }
        '403': { $ref: './components.v1.yaml#/components/responses/Forbidden' }
        '404': { $ref: './components.v1.yaml#/components/responses/NotFound' }
        '409': { $ref: './components.v1.yaml#/components/responses/Conflict' }
  /v1/admin/instances/{instanceId}/purges:
    parameters:
      - $ref: './components.v1.yaml#/components/parameters/InstanceId'
    post:
      operationId: purgeInstance
      tags: [Administration]
      summary: Accept an explicit guarded purge command
      x-audience: administrator
      x-idempotency: required
      x-max-body-bytes: 65536
      x-error-codes:
        [
          ADMIN_REQUIRED,
          INSTANCE_NOT_FOUND,
          RETENTION_NOT_EXPIRED,
          INSTANCE_BUSY,
          OWNERSHIP_NOT_PROVEN,
          IDEMPOTENCY_CONFLICT,
        ]
      x-requirements:
        [SRS-FR-004, SRS-FR-005, SRS-FR-053, SRS-FR-054, SRS-FR-055, SRS-FR-056, SRS-NFR-019]
      parameters:
        - $ref: './components.v1.yaml#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: './components.v1.yaml#/components/schemas/PurgeRequest' }
      responses:
        '202': { $ref: './components.v1.yaml#/components/responses/MutationAccepted' }
        '401': { $ref: './components.v1.yaml#/components/responses/Unauthorized' }
        '403': { $ref: './components.v1.yaml#/components/responses/Forbidden' }
        '404': { $ref: './components.v1.yaml#/components/responses/NotFound' }
        '409': { $ref: './components.v1.yaml#/components/responses/Conflict' }
        '422': { $ref: './components.v1.yaml#/components/responses/Unprocessable' }
  /v1/admin/instances/{instanceId}/reconciliations:
    parameters:
      - $ref: './components.v1.yaml#/components/parameters/InstanceId'
    post:
      operationId: requestReconciliation
      tags: [Administration]
      summary: Request a non-destructive provider observation
      x-audience: administrator
      x-idempotency: required
      x-max-body-bytes: 65536
      x-error-codes: [ADMIN_REQUIRED, INSTANCE_NOT_FOUND, IDEMPOTENCY_CONFLICT]
      x-requirements: [SRS-FR-005, SRS-FR-069, SRS-FR-071, SRS-FR-073]
      parameters:
        - $ref: './components.v1.yaml#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: './components.v1.yaml#/components/schemas/ReconciliationRequest' }
      responses:
        '202': { $ref: './components.v1.yaml#/components/responses/MutationAccepted' }
        '401': { $ref: './components.v1.yaml#/components/responses/Unauthorized' }
        '403': { $ref: './components.v1.yaml#/components/responses/Forbidden' }
        '404': { $ref: './components.v1.yaml#/components/responses/NotFound' }
        '409': { $ref: './components.v1.yaml#/components/responses/Conflict' }
...
  /v1/admin/audit-events:
    get:
      operationId: listAuditEvents
      tags: [Administration]
      summary: List attributed audit events
      x-audience: administrator
      x-idempotency: read-only
      x-error-codes: [VALIDATION_FAILED, ADMIN_REQUIRED]
      x-requirements: [SRS-FR-005, SRS-FR-089, SRS-NFR-019]
      parameters:
        - $ref: './components.v1.yaml#/components/parameters/Cursor'
        - $ref: './components.v1.yaml#/components/parameters/Limit'
        - name: projectId
          in: query
          schema: { type: string, format: uuid }
        - name: operationId
          in: query
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Audit-event page
          content:
            application/json:
              schema: { $ref: './components.v1.yaml#/components/schemas/AuditEventPage' }
        '400': { $ref: './components.v1.yaml#/components/responses/BadRequest' }
        '401': { $ref: './components.v1.yaml#/components/responses/Unauthorized' }
        '403': { $ref: './components.v1.yaml#/components/responses/Forbidden' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT